refactor(logs): move client-side logic to separate component
This commit extracts the client-side logic from the logs page into a dedicated `page.client.tsx` component. This improves maintainability by separating concerns and making the codebase more modular.
